app/requirements.txt
aiodns==3.0.0
aiohttp~=3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
bcrypt==4.0.1
Brotli==1.0.9
cffi==1.15.1
charset-normalizer==2.1.1
frozenlist==1.3.3
greenlet==2.0.1
gunicorn==20.1.0
multidict==6.0.2
pycares==4.2.2
pycparser==2.21
SQLAlchemy==1.4.44
yarl==1.8.1
black==22.10.0
pytest==7.2.0
pydotenv==0.0.7
psycopg2-binary==2.9.5
requests==2.28.1
SQLAlchemy==1.4.44